Moving tests to lib
